pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Event Processor applications"
readme = "README.md"
license-files = ["LICENSE"]
authors = [
    { name = "Tai Lucas", email = "tglucas@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "awscli>=1.42.30",
    "fastapi>=0.116.1",
    "flask>=3.1.2",
    "flask-compress>=1.18",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "permit>=2.8.1",
    "pydantic>=2.11.9",
    "pylru>=1.2.1",
    "python-telegram-bot>=22.4",
    "pytz>=2025.2",
    "pyzmq>=27.0.2",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sentry-sdk[flask]>=2.37.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "tailucas-pylib>=0.5.2",
    "uvicorn[standard]>=0.35.0",
]

[build-system]
requires = ["uv_build>=0.8.14,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "app"
module-root = ""

[dependency-groups]
dev = [
    "pytest>=8.4.2",
]

[project.scripts]
aws_configure = "tailucas_pylib.tools.aws_configure:main"
config_interpol = "tailucas_pylib.tools.config_interpol:main"
cred_tool = "tailucas_pylib.tools.cred_tool:main"
yaml_interpol = "tailucas_pylib.tools.yaml_interpol:main"
